[airbnb] add config option 'ssh_use_agent'
When enabled, this config option turns off using the Vagrant-supplied
identity file, so that unison's ssh command will use the system's
ssh-agent or ssh config.

One reason you might want to do this is to use a PKCS11 hardware token
to connect to your vagrant machine. Vagrant itself is lagging on support
for this feature; at Airbnb we use a monkey-patch.